refactor(blazor): extract IRenderer strategy and drop SKBlazor prefix from internals
Encapsulate the per-host 'if bridged else wasm' branching that lived in
SKCanvasView/SKGLView behind an internal IRenderer abstraction with three
implementations:

- CanvasDirectRenderer / GLDirectRenderer (Blazor WebAssembly, browser-only)
- BridgedRenderer (Blazor Server / Hybrid / static SSR)

The components now hold a single IRenderer, pick the right one in
OnAfterRenderAsync, and delegate Invalidate/Dispose/Dpi and parameter
forwarding through the interface. A unified RenderPaintHandler delegate lets
the bridged renderer own the SKGLView placeholder GL render-target.

Rename internals to drop the redundant SKBlazor prefix (SKBlazorHost -> Host,
SKBlazorHostKind -> HostKind, SKBlazorFrameProducer -> FrameProducer,
SKBlazorBridgedRenderer -> BridgedRenderer) and use file-scoped namespaces for
all new/updated internal and public-API files. Public API names keep the
SKBlazor prefix. Tests and the conformance map are updated to match.
