+++ source/SkiaSharp.Views/SkiaSharp.Views.Blazor/SKCanvasView.razor.cs
@@ -1,6 +1,5 @@
-﻿using System;
+using System;
 using System.Collections.Generic;
-using System.Runtime.InteropServices;
 using System.Runtime.Versioning;
 using System.Threading.Tasks;
 using Microsoft.AspNetCore.Components;
@@ -14,17 +13,9 @@ namespace SkiaSharp.Views.Blazor
 #endif
 	public partial class SKCanvasView : IDisposable
 	{
-		private SKHtmlCanvasInterop interop = null!;
-		private SizeWatcherInterop sizeWatcher = null!;
-		private DpiWatcherInterop dpiWatcher = null!;
 		private ElementReference htmlCanvas;
-
-		private SKSizeI pixelSize;
-		private byte[]? pixels;
-		private GCHandle pixelsHandle;
+		private IRenderer? renderer;
 		private bool ignorePixelScaling;
-		private double dpi;
-		private SKSize canvasSize;
 		private bool enableRenderLoop;
 
 		[Inject]
@@ -33,10 +24,6 @@ namespace SkiaSharp.Views.Blazor
 #if NET9_0_OR_GREATER
 		[Inject]
 		IServiceProvider Services { get; set; } = null!;
-
-		private SKBlazorHostKind hostKind;
-		private bool isBridged;
-		private SKBlazorBridgedRenderer? renderer;
 
 		/// <summary>
 		/// The frame transfer format to use when this view runs in a bridged host (Blazor Server,
@@ -89,17 +76,7 @@ namespace SkiaSharp.Views.Blazor
 		[Parameter(CaptureUnmatchedValues = true)]
 		public IReadOnlyDictionary<string, object>? AdditionalAttributes { get; set; }
 
-		public double Dpi
-		{
-			get
-			{
-#if NET9_0_OR_GREATER
-				if (isBridged && renderer != null)
-					return renderer.Dpi;
-#endif
-				return dpi;
-			}
-		}
+		public double Dpi => renderer?.Dpi ?? 0;
 
 		protected override async Task OnAfterRenderAsync(bool firstRender)
 		{
@@ -107,167 +84,59 @@ namespace SkiaSharp.Views.Blazor
 				return;
 
 #if NET9_0_OR_GREATER
-			hostKind = SKBlazorHost.Resolve(RendererInfo.Name);
-			isBridged = SKBlazorHost.IsBridged(hostKind);
-			if (isBridged)
+			var hostKind = Host.Resolve(RendererInfo.Name);
+			if (Host.IsBridged(hostKind))
 			{
 				var options = (Services?.GetService(typeof(SKBlazorOptions)) as SKBlazorOptions) ?? SKBlazorOptions.Default;
-				renderer = new SKBlazorBridgedRenderer(JS, isGL: false, hostKind, options, PaintBridgedFrame)
-				{
-					EnableRenderLoop = enableRenderLoop,
-					IgnorePixelScaling = ignorePixelScaling,
-					TransferFormat = TransferFormat,
-					Quality = Quality,
-				};
-				await renderer.InitializeAsync(htmlCanvas);
-				return;
+				renderer = new BridgedRenderer(JS, isGL: false, hostKind, options, PaintFrame);
 			}
+			else if (OperatingSystem.IsBrowser())
+			{
+				renderer = new CanvasDirectRenderer(JS, PaintFrame);
+			}
+#else
+			renderer = new CanvasDirectRenderer(JS, PaintFrame);
 #endif
 
-			if (OperatingSystem.IsBrowser())
-			{
-				interop = await SKHtmlCanvasInterop.ImportAsync(JS, htmlCanvas, OnRenderFrame);
-				interop.InitRaster();
+			if (renderer == null)
+				return;
 
-				sizeWatcher = await SizeWatcherInterop.ImportAsync(JS, htmlCanvas, OnSizeChanged);
-				dpiWatcher = await DpiWatcherInterop.ImportAsync(JS, OnDpiChanged);
-			}
+			SyncRenderer();
+			await renderer.InitializeAsync(htmlCanvas);
 		}
 
 		public void Invalidate()
 		{
+			if (renderer == null)
+				return;
+
+			SyncRenderer();
+			renderer.Invalidate();
+		}
+
+		private void SyncRenderer()
+		{
+			if (renderer == null)
+				return;
+
+			renderer.EnableRenderLoop = enableRenderLoop;
+			renderer.IgnorePixelScaling = ignorePixelScaling;
 #if NET9_0_OR_GREATER
-			if (isBridged)
-			{
-				if (renderer != null)
-				{
-					renderer.EnableRenderLoop = enableRenderLoop;
-					renderer.IgnorePixelScaling = ignorePixelScaling;
-					renderer.TransferFormat = TransferFormat;
-					renderer.Quality = Quality;
-					renderer.Invalidate();
-				}
-				return;
-			}
+			renderer.TransferFormat = TransferFormat;
+			renderer.Quality = Quality;
 #endif
-
-			if (!OperatingSystem.IsBrowser())
-				return;
-
-			if (canvasSize.Width <= 0 || canvasSize.Height <= 0 || dpi <= 0)
-				return;
-
-			interop.RequestAnimationFrame(EnableRenderLoop, (int)(canvasSize.Width * dpi), (int)(canvasSize.Height * dpi));
 		}
 
-#if NET9_0_OR_GREATER
-		private void PaintBridgedFrame(SKSurface surface, SKImageInfo userInfo, SKImageInfo rawInfo) =>
-			OnPaintSurface?.Invoke(new SKPaintSurfaceEventArgs(surface, userInfo, rawInfo));
-#endif
-
-		[SupportedOSPlatform("browser")]
-		private void OnRenderFrame()
-		{
-			if (canvasSize.Width <= 0 || canvasSize.Height <= 0 || dpi <= 0)
-				return;
-
-			var info = CreateBitmap(out var unscaledSize);
-			var userVisibleSize = IgnorePixelScaling ? unscaledSize : info.Size;
-
-			using (var surface = SKSurface.Create(info, pixelsHandle.AddrOfPinnedObject(), info.RowBytes))
-			{
-				if (IgnorePixelScaling)
-				{
-					var canvas = surface.Canvas;
-					canvas.Scale((float)dpi);
-					canvas.Save();
-				}
-
-				OnPaintSurface?.Invoke(new SKPaintSurfaceEventArgs(surface, info.WithSize(userVisibleSize), info));
-			}
-
-			interop.PutImageData(pixelsHandle.AddrOfPinnedObject(), info.Size);
-		}
-
-		private SKImageInfo CreateBitmap(out SKSizeI unscaledSize)
-		{
-			var size = CreateSize(out unscaledSize);
-			var info = new SKImageInfo(size.Width, size.Height, SKImageInfo.PlatformColorType, SKAlphaType.Opaque);
-
-			if (pixels == null || pixelSize.Width != info.Width || pixelSize.Height != info.Height)
-			{
-				FreeBitmap();
-
-				pixels = new byte[info.BytesSize];
-				pixelsHandle = GCHandle.Alloc(pixels, GCHandleType.Pinned);
-				pixelSize = info.Size;
-			}
-
-			return info;
-		}
-
-		private SKSizeI CreateSize(out SKSizeI unscaledSize)
-		{
-			unscaledSize = SKSizeI.Empty;
-
-			var w = canvasSize.Width;
-			var h = canvasSize.Height;
-
-			if (!IsPositive(w) || !IsPositive(h))
-				return SKSizeI.Empty;
-
-			unscaledSize = new SKSizeI((int)w, (int)h);
-			return new SKSizeI((int)(w * dpi), (int)(h * dpi));
-
-			static bool IsPositive(double value)
-			{
-				return !double.IsNaN(value) && !double.IsInfinity(value) && value > 0;
-			}
-		}
-
-		private void FreeBitmap()
-		{
-			if (pixels != null)
-			{
-				pixelsHandle.Free();
-				pixels = null;
-			}
-		}
-
-		private void OnDpiChanged(double newDpi)
-		{
-			dpi = newDpi;
-
-			Invalidate();
-		}
-
-		private void OnSizeChanged(SKSize newSize)
-		{
-			if ((int)(canvasSize.Width * dpi) == newSize.Width && (int)(canvasSize.Height * dpi) == newSize.Height)
-				return;
-			canvasSize = newSize;
-
-			Invalidate();
-		}
+		private void PaintFrame(SKSurface surface, SKImageInfo info, SKImageInfo rawInfo, GRBackendRenderTarget? glRenderTarget, GRSurfaceOrigin glOrigin) =>
+			OnPaintSurface?.Invoke(new SKPaintSurfaceEventArgs(surface, info, rawInfo));
 
 		public void Dispose()
 		{
-#if NET9_0_OR_GREATER
 			if (renderer != null)
 			{
 				_ = renderer.DisposeAsync();
 				renderer = null;
 			}
-#endif
-
-			if (OperatingSystem.IsBrowser())
-			{
-				dpiWatcher?.Unsubscribe(OnDpiChanged);
-				sizeWatcher?.Dispose();
-				interop?.Dispose();
-			}
-
-			FreeBitmap();
 		}
 	}
 }
