+++ source/SkiaSharp.Views/SkiaSharp.Views.Blazor/SKGLView.razor.cs
@@ -1,4 +1,4 @@
-﻿using System;
+using System;
 using System.Collections.Generic;
 using System.Runtime.Versioning;
 using System.Threading.Tasks;
@@ -13,26 +13,10 @@ namespace SkiaSharp.Views.Blazor
 #endif
 	public partial class SKGLView : IDisposable
 	{
-		private SKHtmlCanvasInterop interop = null!;
-		private SizeWatcherInterop sizeWatcher = null!;
-		private DpiWatcherInterop dpiWatcher = null!;
-		private SKHtmlCanvasInterop.GLInfo jsGLInfo = null!;
 		private ElementReference htmlCanvas;
-
-		private const int ResourceCacheBytes = 256 * 1024 * 1024; // 256 MB
-		private const SKColorType colorType = SKColorType.Rgba8888;
-		private const GRSurfaceOrigin surfaceOrigin = GRSurfaceOrigin.BottomLeft;
-
-		private GRContext? context;
-		private GRGlInterface? glInterface;
-		private GRBackendRenderTarget? renderTarget;
-		private SKSize renderTargetSize;
-		private SKSurface? surface;
-		private SKCanvas? canvas;
-		private bool enableRenderLoop;
+		private IRenderer? renderer;
 		private bool ignorePixelScaling;
-		private double dpi;
-		private SKSize canvasSize;
+		private bool enableRenderLoop;
 
 		[Inject]
 		IJSRuntime JS { get; set; } = null!;
@@ -40,10 +24,6 @@ namespace SkiaSharp.Views.Blazor
 #if NET9_0_OR_GREATER
 		[Inject]
 		IServiceProvider Services { get; set; } = null!;
-
-		private SKBlazorHostKind hostKind;
-		private bool isBridged;
-		private SKBlazorBridgedRenderer? renderer;
 
 		/// <summary>
 		/// The frame transfer format to use when this view runs in a bridged host (Blazor Server,
@@ -97,17 +77,7 @@ namespace SkiaSharp.Views.Blazor
 		[Parameter(CaptureUnmatchedValues = true)]
 		public IReadOnlyDictionary<string, object>? AdditionalAttributes { get; set; }
 
-		public double Dpi
-		{
-			get
-			{
-#if NET9_0_OR_GREATER
-				if (isBridged && renderer != null)
-					return renderer.Dpi;
-#endif
-				return dpi;
-			}
-		}
+		public double Dpi => renderer?.Dpi ?? 0;
 
 		protected override async Task OnAfterRenderAsync(bool firstRender)
 		{
@@ -115,187 +85,58 @@ namespace SkiaSharp.Views.Blazor
 				return;
 
 #if NET9_0_OR_GREATER
-			hostKind = SKBlazorHost.Resolve(RendererInfo.Name);
-			isBridged = SKBlazorHost.IsBridged(hostKind);
-			if (isBridged)
+			var hostKind = Host.Resolve(RendererInfo.Name);
+			if (Host.IsBridged(hostKind))
 			{
 				var options = (Services?.GetService(typeof(SKBlazorOptions)) as SKBlazorOptions) ?? SKBlazorOptions.Default;
-				renderer = new SKBlazorBridgedRenderer(JS, isGL: true, hostKind, options, PaintBridgedFrame)
-				{
-					EnableRenderLoop = enableRenderLoop,
-					IgnorePixelScaling = ignorePixelScaling,
-					TransferFormat = TransferFormat,
-					Quality = Quality,
-				};
-				await renderer.InitializeAsync(htmlCanvas);
-				return;
+				renderer = new BridgedRenderer(JS, isGL: true, hostKind, options, PaintFrame);
 			}
+			else if (OperatingSystem.IsBrowser())
+			{
+				renderer = new GLDirectRenderer(JS, PaintFrame);
+			}
+#else
+			renderer = new GLDirectRenderer(JS, PaintFrame);
 #endif
 
-			if (OperatingSystem.IsBrowser())
-			{
-				interop = await SKHtmlCanvasInterop.ImportAsync(JS, htmlCanvas, OnRenderFrame);
-				jsGLInfo = interop.InitGL();
+			if (renderer == null)
+				return;
 
-				sizeWatcher = await SizeWatcherInterop.ImportAsync(JS, htmlCanvas, OnSizeChanged);
-				dpiWatcher = await DpiWatcherInterop.ImportAsync(JS, OnDpiChanged);
-			}
+			SyncRenderer();
+			await renderer.InitializeAsync(htmlCanvas);
 		}
 
 		public void Invalidate()
 		{
+			if (renderer == null)
+				return;
+
+			SyncRenderer();
+			renderer.Invalidate();
+		}
+
+		private void SyncRenderer()
+		{
+			if (renderer == null)
+				return;
+
+			renderer.EnableRenderLoop = enableRenderLoop;
+			renderer.IgnorePixelScaling = ignorePixelScaling;
 #if NET9_0_OR_GREATER
-			if (isBridged)
-			{
-				if (renderer != null)
-				{
-					renderer.EnableRenderLoop = enableRenderLoop;
-					renderer.IgnorePixelScaling = ignorePixelScaling;
-					renderer.TransferFormat = TransferFormat;
-					renderer.Quality = Quality;
-					renderer.Invalidate();
-				}
-				return;
-			}
+			renderer.TransferFormat = TransferFormat;
+			renderer.Quality = Quality;
 #endif
-
-			if (!OperatingSystem.IsBrowser())
-				return;
-
-			if (canvasSize.Width <= 0 || canvasSize.Height <= 0 || dpi <= 0 || jsGLInfo == null)
-				return;
-
-			interop.RequestAnimationFrame(EnableRenderLoop, (int)(canvasSize.Width * dpi), (int)(canvasSize.Height * dpi));
 		}
 
-#if NET9_0_OR_GREATER
-		private void PaintBridgedFrame(SKSurface surface, SKImageInfo userInfo, SKImageInfo rawInfo)
-		{
-			// Server-side drawing is CPU raster; present into the WebGL-backed canvas. The render
-			// target is a lightweight placeholder so the GL event args shape is preserved.
-			using var placeholder = new GRBackendRenderTarget(
-				rawInfo.Width,
-				rawInfo.Height,
-				0,
-				0,
-				new GRGlFramebufferInfo(0, rawInfo.ColorType.ToGlSizedFormat()));
-
-			OnPaintSurface?.Invoke(new SKPaintGLSurfaceEventArgs(surface, placeholder, GRSurfaceOrigin.TopLeft, userInfo, rawInfo));
-		}
-#endif
-
-		[SupportedOSPlatform("browser")]
-		private void OnRenderFrame()
-		{
-			if (canvasSize.Width <= 0 || canvasSize.Height <= 0 || dpi <= 0 || jsGLInfo == null)
-				return;
-
-			// create the SkiaSharp context
-			if (context == null)
-			{
-				glInterface = GRGlInterface.Create();
-				context = GRContext.CreateGl(glInterface);
-
-				// bump the default resource cache limit
-				context.SetResourceCacheLimit(ResourceCacheBytes);
-			}
-
-			// get the new surface size
-			var newSize = CreateSize(out var unscaledSize);
-			var info = new SKImageInfo(newSize.Width, newSize.Height, colorType);
-			var userVisibleSize = IgnorePixelScaling ? unscaledSize : info.Size;
-
-			// manage the drawing surface
-			if (renderTarget == null || renderTargetSize != newSize || !renderTarget.IsValid)
-			{
-				// create or update the dimensions
-				renderTargetSize = newSize;
-
-				var glInfo = new GRGlFramebufferInfo(jsGLInfo.FboId, colorType.ToGlSizedFormat());
-
-				// destroy the old surface
-				surface?.Dispose();
-				surface = null;
-				canvas = null;
-
-				// re-create the render target
-				renderTarget?.Dispose();
-				renderTarget = new GRBackendRenderTarget(newSize.Width, newSize.Height, jsGLInfo.Samples, jsGLInfo.Stencils, glInfo);
-			}
-
-			// create the surface
-			if (surface == null)
-			{
-				surface = SKSurface.Create(context, renderTarget, surfaceOrigin, colorType);
-				canvas = surface.Canvas;
-			}
-
-			using (new SKAutoCanvasRestore(canvas, true))
-			{
-				if (IgnorePixelScaling)
-				{
-					var canvas = surface.Canvas;
-					canvas.Scale((float)dpi);
-					canvas.Save();
-				}
-
-				// start drawing
-				OnPaintSurface?.Invoke(new SKPaintGLSurfaceEventArgs(surface, renderTarget, surfaceOrigin, info.WithSize(userVisibleSize), info));
-			}
-
-			// update the control
-			canvas?.Flush();
-			context.Flush();
-		}
-
-		private void OnDpiChanged(double newDpi)
-		{
-			dpi = newDpi;
-
-			Invalidate();
-		}
-
-		private void OnSizeChanged(SKSize newSize)
-		{
-			canvasSize = newSize;
-
-			Invalidate();
-		}
-
-		private SKSizeI CreateSize(out SKSizeI unscaledSize)
-		{
-			unscaledSize = SKSizeI.Empty;
-
-			var w = canvasSize.Width;
-			var h = canvasSize.Height;
-
-			if (!IsPositive(w) || !IsPositive(h))
-				return SKSizeI.Empty;
-
-			unscaledSize = new SKSizeI((int)w, (int)h);
-			return new SKSizeI((int)(w * dpi), (int)(h * dpi));
-
-			static bool IsPositive(double value)
-			{
-				return !double.IsNaN(value) && !double.IsInfinity(value) && value > 0;
-			}
-		}
+		private void PaintFrame(SKSurface surface, SKImageInfo info, SKImageInfo rawInfo, GRBackendRenderTarget? glRenderTarget, GRSurfaceOrigin glOrigin) =>
+			OnPaintSurface?.Invoke(new SKPaintGLSurfaceEventArgs(surface, glRenderTarget!, glOrigin, info, rawInfo));
 
 		public void Dispose()
 		{
-#if NET9_0_OR_GREATER
 			if (renderer != null)
 			{
 				_ = renderer.DisposeAsync();
 				renderer = null;
-			}
-#endif
-
-			if (OperatingSystem.IsBrowser())
-			{
-				dpiWatcher?.Unsubscribe(OnDpiChanged);
-				sizeWatcher?.Dispose();
-				interop?.Dispose();
 			}
 		}
 	}
