+++ tests/SkiaSharp.Tests.Blazor/FrameProducerTests.cs
@@ -5,11 +5,11 @@ using Xunit;
 
 namespace SkiaSharp.Tests.Blazor;
 
-public class SKBlazorFrameProducerTests
+public class FrameProducerTests
 {
 	private static SKImage CreateImage(int width = 8, int height = 8, SKColor? color = null)
 	{
-		var info = new SKImageInfo(width, height, SKBlazorFrameProducer.RgbaColorType, SKAlphaType.Premul);
+		var info = new SKImageInfo(width, height, FrameProducer.RgbaColorType, SKAlphaType.Premul);
 		using var surface = SKSurface.Create(info);
 		surface.Canvas.Clear(color ?? SKColors.Red);
 		return surface.Snapshot();
@@ -20,7 +20,7 @@ public class SKBlazorFrameProducerTests
 	{
 		using var image = CreateImage();
 
-		var bytes = SKBlazorFrameProducer.Produce(image, SKBlazorTransferFormat.Png, 100);
+		var bytes = FrameProducer.Produce(image, SKBlazorTransferFormat.Png, 100);
 
 		Assert.True(bytes.Length > 8);
 		// PNG signature: 89 50 4E 47
@@ -28,7 +28,7 @@ public class SKBlazorFrameProducerTests
 		Assert.Equal(0x50, bytes[1]);
 		Assert.Equal(0x4E, bytes[2]);
 		Assert.Equal(0x47, bytes[3]);
-		Assert.Equal("image/png", SKBlazorFrameProducer.GetContentType(SKBlazorTransferFormat.Png));
+		Assert.Equal("image/png", FrameProducer.GetContentType(SKBlazorTransferFormat.Png));
 	}
 
 	[Fact]
@@ -36,13 +36,13 @@ public class SKBlazorFrameProducerTests
 	{
 		using var image = CreateImage();
 
-		var bytes = SKBlazorFrameProducer.Produce(image, SKBlazorTransferFormat.Jpeg, 80);
+		var bytes = FrameProducer.Produce(image, SKBlazorTransferFormat.Jpeg, 80);
 
 		Assert.True(bytes.Length > 3);
 		// JPEG start-of-image marker: FF D8
 		Assert.Equal(0xFF, bytes[0]);
 		Assert.Equal(0xD8, bytes[1]);
-		Assert.Equal("image/jpeg", SKBlazorFrameProducer.GetContentType(SKBlazorTransferFormat.Jpeg));
+		Assert.Equal("image/jpeg", FrameProducer.GetContentType(SKBlazorTransferFormat.Jpeg));
 	}
 
 	[Fact]
@@ -50,7 +50,7 @@ public class SKBlazorFrameProducerTests
 	{
 		using var image = CreateImage(4, 3, SKColors.Red);
 
-		var bytes = SKBlazorFrameProducer.Produce(image, SKBlazorTransferFormat.Put, 0);
+		var bytes = FrameProducer.Produce(image, SKBlazorTransferFormat.Put, 0);
 
 		Assert.Equal(4 * 3 * 4, bytes.Length);
 		// first pixel is opaque red in RGBA order (directly usable by ImageData/texImage2D)
@@ -58,7 +58,7 @@ public class SKBlazorFrameProducerTests
 		Assert.Equal(0, bytes[1]);
 		Assert.Equal(0, bytes[2]);
 		Assert.Equal(255, bytes[3]);
-		Assert.Equal("application/octet-stream", SKBlazorFrameProducer.GetContentType(SKBlazorTransferFormat.Put));
+		Assert.Equal("application/octet-stream", FrameProducer.GetContentType(SKBlazorTransferFormat.Put));
 	}
 
 	[Fact]
@@ -66,7 +66,7 @@ public class SKBlazorFrameProducerTests
 	{
 		using var image = CreateImage(2, 2, SKColors.Transparent);
 
-		var bytes = SKBlazorFrameProducer.Produce(image, SKBlazorTransferFormat.Put, 0);
+		var bytes = FrameProducer.Produce(image, SKBlazorTransferFormat.Put, 0);
 
 		// fully transparent => alpha 0
 		Assert.Equal(0, bytes[3]);
@@ -80,7 +80,7 @@ public class SKBlazorFrameProducerTests
 	{
 		using var image = CreateImage();
 
-		var bytes = SKBlazorFrameProducer.Produce(image, SKBlazorTransferFormat.Jpeg, quality);
+		var bytes = FrameProducer.Produce(image, SKBlazorTransferFormat.Jpeg, quality);
 
 		Assert.NotEmpty(bytes);
 	}
